chore: create BodyWeightLogController
